Develop page objects for UPS shipment creation pages, sections and forms
Homework 15 (01/20/2022)
Created UpsBasePage class that other pages and page parts extend. Decided that composition "has-a" relationship like that sections are parts of page and form is part of section are applicable for USP site page object design.
TO DO: Parameterize form class based on type: origin, return address or destination for "unguided" shipment creation when all forms are on the same page.